requirements.txt
docker-py==1.2.3
requests==2.7.0
six==1.9.0